fixed : comment, clean and tidy the entire code (use import/export)
