+++ assets/scss/_home.scss
@@ -69,3 +69,8 @@ article{
   color: #636e72;
   padding: 1rem;
 }
+
+.btn{
+  user-select: auto;
+  cursor: auto;
+}
